Verified correct installation of plexapi in Travis
The missed dependency to the 'mock' package on Python 2.7 made it clear
that this went by undetected in the Travis runs.

This change updates the Travis control file to verify that the plexapi
package can be successfully imported after being installed from the
working directory using pip.

Because the installation of py.test drags in a lot of dependencies,
it was moved after the installation of plexapi.

Signed-off-by: Andreas Maier <andreas.r.maier@gmx.de>