change match paramters to be optinally set
Both parameters are not required to be set.
By default, auto macht via gui sends only the agent and language parameters.

The search on the other hand sends both title and year, even if not set.
Yet both parameters are optional, the match will complete its work if the year parameter is not present. And if the title is missing, it basically becomes an ""Auto Match"